pyproject.toml
[project]
name = "alfred-cli"
version = "0.1.0"
description = "Modular AI chat agent with rich terminal interface"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.30.0",
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyperclip>=1.8.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
]

[project.scripts]
alfred = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["alfred"]

[tool.ruff]
line-length = 100
target-version = "py312"
exclude = ["examplecode/"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
